+++ extension/src/Template/TemplatePanel.ts
@@ -46,6 +46,7 @@ export class TemplateEditorPanel {
       {
         // Enable javascript in the webview
         enableScripts: true,
+        retainContextWhenHidden: true,
 
         // And restrict the webview to only loading content from our extension's `TemplateEditor` frontend directory.
         localResourceRoots: [